build release 1.0

+++ arch_ex_button_led_stm32/Adpters/digital_io_adapter.h
@@ -21,7 +21,7 @@ typedef void*          variant_t;
 /**
  * @brief Type definition for pin number.
  */
-typedef uint8_t        pin_t;
+typedef uint16_t        pin_t;
 
 /**
  * @brief Type definition for port.
